+++ horovod/tensorflow/keras/callbacks.py
@@ -82,12 +82,13 @@ class LearningRateScheduleCallback(_impl.LearningRateScheduleCallbackImpl, keras
     `initial_lr` is the learning rate of the model optimizer at the start of the training.
     """
 
-    def __init__(self, multiplier, start_epoch=0, end_epoch=None, staircase=True,
-                 momentum_correction=True, steps_per_epoch=None, initial_lr=None):
+    def __init__(self, initial_lr, multiplier, start_epoch=0, end_epoch=None, staircase=True,
+                 momentum_correction=True, steps_per_epoch=None):
         """
         Construct a new LearningRateScheduleCallback.
 
         Args:
+            initial_lr: Initial learning rate at the start of training.
             multiplier: A constant multiplier or a function `f(epoch) = lr'`
             start_epoch: The first epoch this adjustment will be applied to. Defaults to 0.
             end_epoch: The epoch this adjustment will stop applying (exclusive end).
@@ -99,14 +100,9 @@ class LearningRateScheduleCallback(_impl.LearningRateScheduleCallbackImpl, keras
             steps_per_epoch: The callback will attempt to autodetect number of batches per
                              epoch with Keras >= 2.0.0. Provide this value if you have an older
                              version of Keras.
-            initial_lr: Initial learning rate at the start of training.
-
-                .. warning:: Will be required in v0.21.0.
-
         """
-        super(LearningRateScheduleCallback, self).__init__(K, multiplier, start_epoch, end_epoch,
-                                                           staircase, momentum_correction, steps_per_epoch,
-                                                           initial_lr)
+        super(LearningRateScheduleCallback, self).__init__(K, initial_lr, multiplier, start_epoch, end_epoch,
+                                                           staircase, momentum_correction, steps_per_epoch)
 
 
 class LearningRateWarmupCallback(_impl.LearningRateWarmupCallbackImpl, keras.callbacks.Callback):
@@ -133,12 +129,13 @@ class LearningRateWarmupCallback(_impl.LearningRateWarmupCallbackImpl, keras.cal
         lr'(epoch = warmup) &= lr
     """
 
-    def __init__(self, warmup_epochs=5, momentum_correction=True, steps_per_epoch=None,
-                 verbose=0, initial_lr=None):
+    def __init__(self, initial_lr, warmup_epochs=5, momentum_correction=True, steps_per_epoch=None,
+                 verbose=0):
         """
         Construct a new LearningRateWarmupCallback that will gradually warm up the learning rate.
 
         Args:
+            initial_lr: Initial learning rate at the start of training.
             warmup_epochs: The number of epochs of the warmup phase. Defaults to 5.
             momentum_correction: Apply momentum correction to optimizers that have momentum.
                                  Defaults to True.
@@ -146,12 +143,9 @@ class LearningRateWarmupCallback(_impl.LearningRateWarmupCallbackImpl, keras.cal
                              epoch with Keras >= 2.0.0. Provide this value if you have an older
                              version of Keras.
             verbose: verbosity mode, 0 or 1.
-            initial_lr: Initial learning rate at the start of training.
-
-                .. warning:: Will be required in v0.21.0.
         """
-        super(LearningRateWarmupCallback, self).__init__(K, warmup_epochs, momentum_correction,
-                                                         steps_per_epoch, verbose, initial_lr)
+        super(LearningRateWarmupCallback, self).__init__(K, initial_lr, warmup_epochs, momentum_correction,
+                                                         steps_per_epoch, verbose)
 
 
 class BestModelCheckpoint(keras.callbacks.ModelCheckpoint):